requirements.txt
paho-mqtt
NooLite_F
